[refactor][m] Renaming lib folder to dependencies according airflow naming conventions
