+++ aircan/dags/api_ckan_load_multiple_nodes.py
@@ -5,10 +5,10 @@ import ast
 from datetime import datetime
 
 # Local imports
-from aircan.lib.hybrid_load import delete_datastore_table, create_datastore_table
-from aircan.lib.api_load import load_resource_via_api
+from aircan.dependencies.hybrid_load import delete_datastore_table, create_datastore_table
+from aircan.dependencies.api_load import load_resource_via_api
 
-from aircan.lib.file_conversion.csv_to_json import convert
+from aircan.dependencies.file_conversion.csv_to_json import convert
 
 # Third-party library imports
 from airflow import DAG
@@ -16,10 +16,10 @@ from airflow.exceptions import AirflowException
 
 from airflow.models import Variable
 from airflow.operators.python_operator import PythonOperator
+from airflow.utils.dates import days_ago
 
 args = {
-    'owner': 'airflow',
-    'start_date': datetime.now(),
+    'start_date': days_ago(0),
     'params': { 
         "resource_id": "res-id-123",
         "schema_fields_array": "['field1', 'field2']", 
@@ -31,8 +31,7 @@ args = {
 dag = DAG(
     dag_id='ckan_api_load_multiple_steps',
     default_args=args,
-    schedule_interval=None,
-    tags=['conversion'],
+    schedule_interval=None
 )
