--- aircan/lib/r2.csv
@@ -1,3 +1,0 @@
-_id,_full_text,FID
-192607,t,2
-192608,t,3
